+++ packages/lib/core/src/config/env.ts
@@ -1,32 +1,15 @@
-import type { APIGuard, CommandConfig, KVNamespace } from './types';
+import type { APIGuardBinding, KVNamespaceBinding, WorkerAIBinding } from './binding';
+import type { AgentUserConfig, AgentUserConfigKey } from './config';
 import loadI18n from '../i18n';
-import {
-    AgentShareConfig,
-    AnthropicConfig,
-    AzureConfig,
-    CohereConfig,
-    DallEConfig,
-    DefineKeys,
-    EnvironmentConfig,
-    GeminiConfig,
-    MistralConfig,
-    OpenAIConfig,
-    WorkersConfig,
-} from './config';
+import { AgentShareConfig, AnthropicConfig, AzureConfig, CohereConfig, DallEConfig, DeepSeekConfig, DefineKeys, EnvironmentConfig, GeminiConfig, GorqConfig, MistralConfig, OpenAIConfig, WorkersConfig } from './config';
 import { ConfigMerger } from './merger';
 import { BUILD_TIMESTAMP, BUILD_VERSION } from './version';
 
-export type AgentUserConfig = Record<string, any> &
-    DefineKeys &
-    AgentShareConfig &
-    OpenAIConfig &
-    DallEConfig &
-    AzureConfig &
-    WorkersConfig &
-    GeminiConfig &
-    MistralConfig &
-    CohereConfig &
-    AnthropicConfig;
+export interface CommandConfig {
+    value: string;
+    description?: string | null;
+    scope?: string[] | null;
+}
 
 function createAgentUserConfig(): AgentUserConfig {
     return Object.assign(
@@ -41,10 +24,16 @@ function createAgentUserConfig(): AgentUserConfig {
         new MistralConfig(),
         new CohereConfig(),
         new AnthropicConfig(),
+        new DeepSeekConfig(),
+        new GorqConfig(),
     );
 }
 
-export const ENV_KEY_MAPPER: Record<string, string> = {
+function fixApiBase(base: string): string {
+    return base.replace(/\/+$/, '');
+}
+
+export const ENV_KEY_MAPPER: Record<string, AgentUserConfigKey> = {
     CHAT_MODEL: 'OPENAI_CHAT_MODEL',
     API_KEY: 'OPENAI_API_KEY',
     WORKERS_AI_MODEL: 'WORKERS_CHAT_MODEL',
@@ -67,8 +56,10 @@ class Environment extends EnvironmentConfig {
     readonly CUSTOM_COMMAND: Record<string, CommandConfig> = {};
     readonly PLUGINS_COMMAND: Record<string, CommandConfig> = {};
 
-    DATABASE: KVNamespace = null as any;
-    API_GUARD: APIGuard | null = null;
+    AI_BINDING: WorkerAIBinding | null = null;
+    API_GUARD: APIGuardBinding | null = null;
+
+    DATABASE: KVNamespaceBinding = null as any;
 
     CUSTOM_MESSAGE_RENDER: CustomMessageRender | null = null;
 
@@ -79,6 +70,7 @@ class Environment extends EnvironmentConfig {
 
     merge(source: any) {
         // 全局对象
+        this.AI_BINDING = source.AI;
         this.DATABASE = source.DATABASE;
         this.API_GUARD = source.API_GUARD;
 
@@ -123,6 +115,7 @@ class Environment extends EnvironmentConfig {
         ]);
         ConfigMerger.merge(this.USER_CONFIG, source);
         this.migrateOldEnv(source);
+        this.fixAgentUserConfigApiBase();
         this.USER_CONFIG.DEFINE_KEYS = [];
         this.I18N = loadI18n(this.LANGUAGE.toLowerCase());
     }
@@ -186,16 +179,35 @@ class Environment extends EnvironmentConfig {
         if (source.AZURE_COMPLETIONS_API && !this.USER_CONFIG.AZURE_CHAT_MODEL) {
             const url = new URL(source.AZURE_COMPLETIONS_API);
             this.USER_CONFIG.AZURE_RESOURCE_NAME = url.hostname.split('.').at(0) || null;
-            this.USER_CONFIG.AZURE_CHAT_MODEL = url.pathname.split('/').at(3) || null;
+            this.USER_CONFIG.AZURE_CHAT_MODEL = url.pathname.split('/').at(3) || 'gpt-4o-mini';
             this.USER_CONFIG.AZURE_API_VERSION = url.searchParams.get('api-version') || '2024-06-01';
         }
         // 兼容旧版 AZURE_DALLE_API
         if (source.AZURE_DALLE_API && !this.USER_CONFIG.AZURE_IMAGE_MODEL) {
             const url = new URL(source.AZURE_DALLE_API);
             this.USER_CONFIG.AZURE_RESOURCE_NAME = url.hostname.split('.').at(0) || null;
-            this.USER_CONFIG.AZURE_IMAGE_MODEL = url.pathname.split('/').at(3) || null;
+            this.USER_CONFIG.AZURE_IMAGE_MODEL = url.pathname.split('/').at(3) || 'dall-e-3';
             this.USER_CONFIG.AZURE_API_VERSION = url.searchParams.get('api-version') || '2024-06-01';
         }
+    }
+
+    private fixAgentUserConfigApiBase() {
+        const keys: AgentUserConfigKey[] = [
+            'OPENAI_API_BASE',
+            'GOOGLE_API_BASE',
+            'MISTRAL_API_BASE',
+            'COHERE_API_BASE',
+            'ANTHROPIC_API_BASE',
+            'DEEPSEEK_API_BASE',
+            'GORQ_API_BASE',
+        ];
+        for (const key of keys) {
+            const base = this.USER_CONFIG[key];
+            if (this.USER_CONFIG[key] && typeof base === 'string') {
+                this.USER_CONFIG[key] = fixApiBase(base) as any;
+            }
+        }
+        this.TELEGRAM_API_DOMAIN = fixApiBase(this.TELEGRAM_API_DOMAIN);
     }
 }