+++ doc/en/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+- v1.10.6
+  - Add `DeepSeek` and `Gorq` independent Agent support
+
+- v1.10.5
+  - Add full support for `Telegram MarkdownV2`
+
 - v1.10.0
   - Switching Models with InlineKeyboards